Extract out downloads logic from Store to specific module
