+++ src/my_models/delete_model_modal.rs
@@ -158,7 +158,7 @@ impl Widget for DeleteModelModal {
     }
 
     fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let downloaded_files = &scope.data.get::<Store>().unwrap().downloaded_files;
+        let downloaded_files = &scope.data.get::<Store>().unwrap().downloads.downloaded_files;
 
         let downloaded_file = downloaded_files
             .iter()
